pyproject.toml
[project]
name = "tech-chain"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-filter (>=25.2,<26.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "coverage (>=7.13.1,<8.0.0)",
    "ruff (>=0.14.14,<0.15.0)"
]

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = [
    ".git",
    "__pycache__",
    "migrations",
    ".venv",
    "venv",
    ".idea",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "F",  # pyflakes
    "I",  # isort
    "B",  # bugbear
]
ignore = [
    "E203",  # black compatibility
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
